+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Alignment issue between checkboxes and their labels (#376)
 -   Scroll on performance legend overlaps on header "Rank X" (#381)
+-   Restore metadata section to task drawers (#384)
 
 ### Changed